feat(frontend): mobile friendly overview
