+++ packages/frontend/components/Borrow/Wrapper.tsx
@@ -13,7 +13,6 @@ import Borrow from "../../components/Borrow/Borrow"
 import Footer from "../../components/Shared/Footer"
 import Header from "../Shared/Header/Header"
 import Overview from "./Overview/Overview"
-import TransactionSummary from "../../components/Borrow/TransactionSummary"
 import { useBorrow } from "../../store/borrow.store"
 import { useEffect, useState } from "react"
 import { useAuth } from "../../store/auth.store"
@@ -149,11 +148,7 @@ function BorrowWrapper({ query }: BorrowWrapperProps) {
               <Borrow isEditing={isEditing} basePosition={basePosition} />
             </Grid>
             <Grid item sm={12} md={7}>
-              {isMobile ? (
-                <TransactionSummary />
-              ) : (
-                <Overview basePosition={basePosition} />
-              )}
+              <Overview basePosition={basePosition} />
             </Grid>
           </Grid>
         )}
